\docType{data}
\name{fertility.gdp}
\alias{fertility.gdp}
\title{Fertility and GDP Data

Table 14.6 from the 4th edition.}
\format{A \code{data.frame} with 38 observations on the following 3 variables,
\describe{
   \item{\code{Nation}}{nation}
   \item{\code{Fertility}}{Fertility rate}
   \item{\code{GDP}}{GDP per capita}
}}
\source{
  \url{http://www.stat.ufl.edu/~aa/social/data.html}
}
\description{
  Fertility and GDP Data

  Table 14.6 from the 4th edition.
}
\keyword{datasets}

