% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/house.selling.price.2.R
\docType{data}
\name{house.selling.price.2}
\alias{house.selling.price.2}
\title{House Selling Price 2 Data}
\format{\Sexpr[stage=build,results=rd]{data(house.selling.price.2); smss:::describe_df(house.selling.price.2)}
\describe{
\item{\code{P}}{selling price}
\item{\code{Ba}}{number of bathrooms}
\item{\code{New}}{whether new (1 = yes, 0 = no)}
}}
\source{
\url{http://www.stat.ufl.edu/~aa/social/data.html}
}
\description{
Selling price of homes in Gainesville, Florida, January 1996. Table 9.4  in 3rd edition.
Data provided by Jane Myers, Coldwell-Banker Realty.
}
\examples{
data(house.selling.price.2)
summary(house.selling.price.2)
}

