% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/income.credit.R
\docType{data}
\name{income.credit}
\alias{income.credit}
\title{Income and Credit Card Possession Data}
\format{\Sexpr[stage=build,results=rd]{data(income.credit); smss:::df_format(income.credit)}}
\source{
\url{http://www.stat.ufl.edu/~aa/social/data.html}
}
\description{
Data on annual income (euros), number of subjects at that income level, and number possessing##'  a travel credit card.
Source: Based on data in "Categorical Data Analysis," Quaderni del Corso Estivo di Statistica e Calcolo delle Probabilita;, n. 4., Istituto di Metodi Quantitativi, Universita; Luigi Bocconi, a cura di R. Piccarreta (1993).
Table 15.1 of 4th edition.
}
\examples{
data(income.credit)
summary(income.credit)
}

