% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/smss-package.R
\name{smss-package}
\alias{smss-package}
\title{Data from Agresti "Statistical Methods for the Social Sciences"}
\description{
Data-only package with the datasets used by the textbook, \emph{Statististical Methods for the Social Sciences}
by Alan Agresti and Barbara Finlay.

\itemize{
\item{Agresti, Alan and Barbara Finlay (2009), \emph{Statistical Methods for the Social Sciences}, 4th ed., Pearson}
\item{Agresti, Alan and Barbara Finlay (1997), \emph{Statistical Methods for the Social Sciences}, 3rd ed., Pearson}
}

The datasets included in this package are those avaialable at
\url{http://www.stat.ufl.edu/~aa/social/data.html}.
}

