\docType{data}
\name{zagat}
\alias{zagat}
\title{Zagat restraunt rating of Italian Restraunts Data}
\format{\Sexpr[stage=build,results=rd]{data(zagat); smss:::df_format(zagat)}}
\source{
  \url{http://www.stat.ufl.edu/~aa/social/data.html}
}
\description{
  Zagat restraunt rating of Italian Restraunts Data.
  Examples 3.59, 3.60, and 9.38 of 4th edition.
}
\examples{
data(zagat)
summary(zagat)
}
\keyword{datasets}

